+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/ExpectedFormHierarchies.swift
@@ -138,6 +138,11 @@ enum ExpectedFormHierarchy {
     enum ShopeePay {
         static var paymentIntent: FormHierarchyNode { emptyForm }
     }
+    // MARK: - QRIS
+
+    enum QRIS {
+        static var paymentIntent: FormHierarchyNode { emptyForm }
+    }
     // MARK: - PAYCO
 
     enum Payco {